+++ src/utils/utils.js
@@ -1,0 +1,13 @@
+const setEvents = (arr = [], {eventName = `click`, cb} = {}) => {
+  const setNodeEvent = (node) => {
+    node.addEventListener(eventName, cb);
+  };
+
+  if (typeof cb === `function`) {
+    return arr.forEach(setNodeEvent);
+  }
+
+  return false;
+};
+
+export {setEvents};
